fix slider

+++ src/main/java/com/app/toeic/slider/controller/SliderController.java
@@ -1,12 +1,16 @@
 package com.app.toeic.slider.controller;
 
+import com.app.toeic.external.response.ResponseVO;
+import com.app.toeic.external.service.FirebaseStorageService;
 import com.app.toeic.slider.model.Slider;
 import com.app.toeic.slider.payload.SliderDTO;
 import com.app.toeic.slider.repo.SliderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.Objects;
 
 @RestController
@@ -15,7 +19,7 @@ import java.util.Objects;
 @RequestMapping("/slider")
 public class SliderController {
     private final SliderRepository sliderRepository;
-
+    private final FirebaseStorageService firebaseStorageService;
     @GetMapping("/all")
     public Object getAllSlider(
             @RequestParam(value = "page", defaultValue = "1") Integer page,
@@ -25,17 +29,19 @@ public class SliderController {
     }
 
     @PostMapping("/add")
-    public Object addSlider(@RequestBody SliderDTO slider) {
+    public Object addSlider(@RequestParam("file") MultipartFile file) throws
+            IOException {
+        if (file == null) return "FAIL";
+        var image = firebaseStorageService.uploadFile(file);
         var lastSlider = sliderRepository
                 .findLastByPosition();
         var position = 1;
         if (lastSlider.isPresent()) {
             position = lastSlider.get().getPosition() + 1;
         }
-
         var newSlider = Slider
                 .builder()
-                .image(slider.getImage())
+                .image(image)
                 .position(position)
                 .build();
         sliderRepository.save(newSlider);
@@ -44,7 +50,19 @@ public class SliderController {
 
     @DeleteMapping("/delete/{sliderId}")
     public Object deleteSlider(@PathVariable("sliderId") Long sliderId) {
+        var slider = sliderRepository
+                .findById(sliderId)
+                .orElse(null);
+        if (slider == null) return "FAIL";
+        var position = slider.getPosition();
         sliderRepository.deleteById(sliderId);
+        // update position of all slider after deleted slider
+        var sliders = sliderRepository
+                .findAllByPositionGreaterThanOrderByPosition(position);
+        for (var s : sliders) {
+            s.setPosition(s.getPosition() - 1);
+            sliderRepository.save(s);
+        }
         return "OK";
     }
 
@@ -70,7 +88,7 @@ public class SliderController {
         var slider = sliderRepository
                 .findById(sliderId)
                 .orElseThrow();
-        Slider nextSlider = null;
+        Slider nextSlider;
         if (action.equals("up")) {
             slider.setPosition(slider.getPosition() - 1);
             nextSlider = sliderRepository
